backend/api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyPDF2==3.0.1
python-dotenv==0.19.0
requests==2.26.0
langchain==0.0.267
openai==0.27.0
werkzeug==2.0.3 
gunicorn==20.1.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz 
flask-mail==0.9.1